openstack/tests/unit/dns/v2/test_service_status.py
# Licensed under the Apache License, Version 2.0 (the "License"); you may
# not use this file except in compliance with the License. You may obtain
# a copy of the License at
#
#      http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
# License for the specific language governing permissions and limitations
# under the License.

from openstack.dns.v2 import service_status as svc_status
from openstack.tests.unit import base


class TestServiceStatus(base.TestCase):
    def test_basic(self):
        sot = svc_status.ServiceStatus()
        self.assertEqual(None, sot.resource_key)
        self.assertEqual('service_statuses', sot.resources_key)
        self.assertEqual('/service_statuses', sot.base_path)
        self.assertTrue(sot.allow_list)
        self.assertFalse(sot.allow_create)
        self.assertTrue(sot.allow_fetch)
        self.assertFalse(sot.allow_commit)
        self.assertFalse(sot.allow_delete)
